edit

+++ lab1_3.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    unsigned long int n;
+    long long int n;
     int a,b;
-    scanf("%u",&n);
+    scanf("%lld",&n);
     a=n%3;
     b=n%11;
     printf("%d %d",a,b);
